Move GenesisTxParser.findGenesisTx() into TxParser
The GenesisTxParser class was unused except for the one static
method findGenesisTx. Move that method into TxParser and remove
GenesisTxParser.
